requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
django-cors-headers==3.10.0
psycopg2-binary==2.9.1
python-dotenv==0.19.2
drf-yasg==1.20.0